requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
python-binance==1.0.19
python-dotenv==1.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas_ta==0.3.14b0
schedule==1.2.1